+++ packages/joker_state/lib/joker_state.dart
@@ -1,3 +1,5 @@
 export 'package:circus_ring/circus_ring.dart';
 
+export 'cue_gate.dart';
+export 'cue_master.dart';
 export 'src/state_management/joker_states.dart';